tui input refactor continued
link selection now mostly implemented and looks the same as before.  extremely barebones search functionality written as a test (also realised it would conflict with n/p for link selection...).  caret has been improved a lot and now supports per-'word' movements and nice deletion with Ctrl-W.  also implemented protocol cycling feature in input prompts.  still facing issues with line lengths and bugs seem to be appearing every now and then on some pages...

+++ tui_input.h
@@ -63,11 +63,11 @@ struct tui_input
     char buffer[TUI_INPUT_BUFFER_MAX];
 
     // The buffer length; this does NOT include the null-termination character
-    size_t buffer_len;
+    int buffer_len;
 
     // Location of input buffer caret.
     // This is in range of 0 to buffer length + 1
-    size_t caret;
+    int caret;
 
     // Current prompt presented to user
     char prompt[TUI_INPUT_PROMPT_MAX];
